Map functions are now pass-by-value
As a convenience to avoid extra variables, trying out if passing by
value and using the rewriter to construct the pointer is cleaner. It
seems that the code has fewer lines, though in some cases the lines
are longer.

The best example of a cleaner source file is in tc_neighbor_sharing.c.

Signed-off-by: Brenden Blanco <bblanco@plumgrid.com>