docs: add the supplementary commits for LRU hashes
This commit addes two kernel commits for the LRU hashes to provide the
more information about LRU hashes. The original commits (29ba732acbee
and 8f8449384ec3) mark the inclusion of LRU hashes, but the commit
messages do not mention the details. The commits 3a08c2fd7634 and
961578b63474 actually implement the LRU list and provide the detailed
information about why LRU hash is created and how it works, and this can
help people to understand LRU hashes more quickly.

Signed-off-by: Gary Lin <glin@suse.com>